regidurias/requirements.txt
Django==2.2.13
django-filter==21.1
django-widget-tweaks==1.4.12
gunicorn==19.7.1
Pillow==6.2.2
psycopg2-binary==2.8.6
pytz==2021.3
sqlparse==0.4.2
